+++ CHANGELOG.md
@@ -28,6 +28,9 @@
   - 'chipinfo' command to display chip information
   - 'apenable' command to enable/disable the Access Point
 
+- Misc
+  - Add build name config variable for debugging and auto-flash support - @tototo31
+
 ### Changed
 
 - Display
@@ -38,6 +41,7 @@
   - Color status bar icons based on their activity
   - S3TWatch: Disable tap-to-wake, use touch interrupt instead.
   - Exiting a view now returns to the previous view instead of the main menu - @tototo31
+  - Add CAPSLOCK shift toggle to keyboard view - @tototo31
 
 - WebUI
   - Minor style tweaks
@@ -52,11 +56,13 @@
   - Dynamically size error popup to content and center on screen
   - Reduce FatFS memory usage on S3TWatch and Cardputer
   - Improve battery reading accuracy on Cardputer
+  - Fix keyboard view touch detection logic - @tototo31
 
 - Commands
   - Add termianl_view_add_text logs to commands missing them
   - Skip pcap flush if mutex is null
   - Fix stop command not stopping GPS task
+  - Fix serial going unresponsive by using 'scanap -stop'
 
 - General
   - Disable and re-enable ESP comm manager UART around GPS usage to avoid driver conflicts